refator: добавление модуля попапа
