+++ js/api.js
@@ -1,36 +1,13 @@
 import { setFilterListener } from './filters-ads.js';
-import { checkCapacity, showCapacityError } from './form-validity.js';
 import { enableFilters } from './lock-form.js';
 import { createMarkerGroup } from './map.js';
+import { openPopup } from './popup.js';
 
 const SERVER_URL = 'https://23.javascript.pages.academy/keksobooking';
 
 const successTemplate = document.querySelector('#success').content.querySelector('.success');
 const errorTemplate = document.querySelector('#error').content.querySelector('.error');
 const dataErrorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
-const adForm = document.querySelector('.ad-form');
-
-const isEsc = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
-
-const closePopup = (evt) => {
-  const popupActive = document.querySelector('.open-popup');
-  if (evt.type === 'click' || isEsc(evt)) {
-    popupActive.remove();
-    document.removeEventListener('keydown', closePopup);
-  }
-};
-
-const addEventPopup = (popup) => {
-  popup.addEventListener('click', closePopup);
-  document.addEventListener('keydown', closePopup);
-};
-
-const openPopup = (messagePopup) => {
-  const popup = messagePopup.cloneNode(true);
-  document.body.append(popup);
-  popup.classList.add('open-popup');
-  addEventPopup(popup);
-};
 
 let defaultData = [];
 const getData = () => {
@@ -72,20 +49,4 @@ const sendData = (onSuccess, body) => {
     });
 };
 
-const setUserFormSubmit = (onSuccess) => {
-  adForm.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    if (checkCapacity()) {
-      const formData = new FormData(evt.target);
-
-      sendData(
-        onSuccess,
-        formData,
-      );
-    } else {
-      showCapacityError();
-    }
-  });
-};
-
-export {getData, setUserFormSubmit, defaultData};
+export {getData, sendData, defaultData};
