+++ js/main.js
@@ -1,7 +1,7 @@
 import {disableForm} from './lock-form.js';
-import {checkValidityForm, resetForm } from './form-validity.js';
+import {checkValidityForm, resetForm, setUserFormSubmit } from './form-validity.js';
 import {initializationMap} from './map.js';
-import {getData, setUserFormSubmit} from './api.js';
+import {getData} from './api.js';
 import {getPreview} from './preview-img.js';
 
 disableForm();
